+++ frontend/src/components/Lobby/SideBar.js
